beta version almost done

+++ src/loader/loader.c
@@ -89,7 +89,7 @@ ssize_t xread(int fd, void *buf, size_t count)
 }
 
 // copy the instructions into the pages
-void copy_into(so_seg_t *segment, int offset, void *pageAddress)
+void copy_into(so_seg_t *segment,size_t  offset, void *pageAddress)
 {
 	ssize_t pageSize = getpagesize();
 	char *buffer = malloc(pageSize * sizeof(char));
